# Conflicts:
#	src/main/java/fr/eseoye/eseoye/action/OnePost.java
#	src/main/java/fr/eseoye/eseoye/servlet/ESEOyeServlet.java
#	src/main/webapp/js/request.js
#	src/main/webapp/jsp/userPanel/Account.jsp

+++ src/main/java/fr/eseoye/eseoye/io/databases/DatabaseCredentials.java
@@ -8,9 +8,9 @@ public class DatabaseCredentials {
 	private DatabaseType type;
 	private String databaseName;
 	
-	public DatabaseCredentials(String url, int port, String username, String password, String databaseName, DatabaseType type) {
+	public DatabaseCredentials(String url, long port, String username, String password, String databaseName, DatabaseType type) {
 		this.url = url;
-		this.port = port;
+		this.port = Long.valueOf(port).intValue();
 		this.username = username;
 		this.password = password;
